Don't supply base fixities when UInfixE is available.
As of template-haskell 2.7.0, template haskell has a UInfixE/UInfixP
constructor. By not supplying those fixities to haskell-src-exts,
you get an AST which uses those unresolved fixities. This makes parseExp
succeed more often, and means you get the "right" result if you decided
to use  parseExp "1+1*2", in a module that had + and * defined with
the opposite fixities they have in base.